Changed tank's animation. Added Dynamic and Static interface for game objects. Added simple Physics engine class.

+++ src/Manager/ResourceManager.cpp
@@ -316,7 +316,7 @@ bool ResourceManager::loadJSONResources(const std::string& JSONpath)
             for(const auto& frame : framesArray)
             {
                 std::string subTexture = frame["subTexture"].GetString();
-                uint64_t duration = frame["duration"].GetUint64();
+                double duration = frame["duration"].GetDouble();
                 pAnimetedSprite->addFrame(subTexture, duration);
             }
         }
